remove redundant type cast

+++ src/GameLogic/PlayerActions/Items/MoveItemAction.cs
@@ -332,7 +332,7 @@ public class MoveItemAction
                 continue; // to make sure that the same item is not blocking itself
             }
 
-            var blockingItem = toStorage.Storage.GetItem((byte)i);
+            var blockingItem = toStorage.Storage.GetItem(i);
             if (blockingItem is null)
             {
                 continue; // no item is blocking the slot
